+++ src/frontend/apps/impress/src/components/Text.tsx
@@ -13,6 +13,7 @@ export interface TextProps extends BoxProps {
     ReactHTML,
     'p' | 'span' | 'div' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
   >;
+  $elipsis?: boolean;
   $weight?: CSSProperties['fontWeight'];
   $textAlign?: CSSProperties['textAlign'];
   // eslint-disable-next-line @typescript-eslint/ban-types
@@ -49,6 +50,9 @@ export const TextStyled = styled(Box)<TextProps>`
   ${({ $theme, $variation }) =>
     `color: var(--c--theme--colors--${$theme}-${$variation});`}
   ${({ $color }) => $color && `color: ${$color};`}
+  ${({ $elipsis }) =>
+    $elipsis &&
+    `white-space: nowrap; overflow: hidden; text-overflow: ellipsis;`}
 `;
 
 export const Text = ({